# 🔐 Advanced Security + Self-Reflection System + Complete Test Resolution