--- a/reasoning_engine_profiling_report.txt
+++ b/reasoning_engine_profiling_report.txt
@@ -1,0 +1,1 @@
+No memory profiles recorded