--- a/src/main.rs
+++ b/src/main.rs
@@ -4,7 +4,7 @@ use std::process;
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
     env_logger::init();
-    if let Err(e) = cli::run().await {
+    if let Err(e) = cli::run_modular().await {
         eprintln!("Error: {}", e);
         process::exit(1);
     }