… Cargo.toml
… README.md
… agent_config.json
… analysis/reflection_system_analysis.md
… anthropic_config.json
… claude_refactoring_analysis.md
… claude_unwrap_fixes.md
… config.yaml
… crates/fluent-agent/src/orchestrator.rs
… crates/fluent-agent/src/profiling/mod.rs
… crates/fluent-agent/src/reflection.rs
… crates/fluent-agent/src/state_manager.rs
… crates/fluent-agent/src/tools/filesystem.rs
… crates/fluent-agent/src/tools/shell.rs
… crates/fluent-agent/src/tools/string_replace_editor.rs
… crates/fluent-agent/src/tools/string_replace_editor_tests.rs
… crates/fluent-agent/src/transport/http.rs
… crates/fluent-cli/src/frogger.rs
… crates/fluent-cli/src/lib.rs
… crates/fluent-core/src/config.rs
… crates/fluent-engines/Cargo.toml
… crates/fluent-engines/src/optimized_openai.rs
… crates/fluent-engines/src/secure_plugin_system.rs
… docs/guides/agent-system.md
… enhanced_reflection_profiling_report.txt
… examples/agent_frogger.rs
… examples/enhanced_reflection_demo.rs
… examples/reflection_demo.rs
… examples/state_management_demo.rs
… flexible_config.json